refactor : 기존 UseCase 파일 삭제 및 디렉토리 이동 #PAR-213
