--- core/domain/src/main/java/online/partyrun/partyrunapplication/core/domain/SendDeclineMatchUseCase.kt
@@ -1,15 +1,0 @@
-package online.partyrun.partyrunapplication.core.domain
-
-import kotlinx.coroutines.flow.Flow
-import online.partyrun.partyrunapplication.core.common.network.ApiResponse
-import online.partyrun.partyrunapplication.core.data.repository.MatchRepository
-import online.partyrun.partyrunapplication.core.model.match.MatchDecisionRequest
-import online.partyrun.partyrunapplication.core.model.match.MatchStatusResult
-import javax.inject.Inject
-
-class SendDeclineMatchUseCase @Inject constructor(
-    private val matchRepository: MatchRepository
-) {
-    suspend operator fun invoke(matchDecisionRequest: MatchDecisionRequest): Flow<ApiResponse<MatchStatusResult>> =
-        matchRepository.declineMatch(matchDecisionRequest)
-}
